# internal/api/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "immerle.AccountResponse": {
                "properties": {
                    "ok": {
                        "example": true,
                        "type": "boolean"
                    },
                    "user": {
                        "properties": {
                            "displayName": {
                                "example": "Bob Marley",
                                "type": "string"
                            },
                            "email": {
                                "example": "bob@example.com",
                                "type": "string"
                            },
                            "id": {
                                "type": "string"
                            },
                            "isAdmin": {
                                "example": false,
                                "type": "boolean"
                            },
                            "username": {
                                "example": "bob",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "immerle.CleanupRunResponse": {
                "properties": {
                    "ok": {
                        "example": true,
                        "type": "boolean"
                    },
                    "removed": {
                        "example": 3,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "immerle.CleanupStatusResponse": {
                "properties": {
                    "enabled": {
                        "example": true,
                        "type": "boolean"
                    },
                    "intervalSeconds": {
                        "example": 21600,
                        "type": "integer"
                    },
                    "maxAgeSeconds": {
                        "example": 2592000,
                        "type": "integer"
                    },
                    "ok": {
                        "example": true,
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "immerle.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "already_initialized",
                        "type": "string"
                    },
                    "ok": {
                        "example": false,
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "contact": {
            "name": "immerle"
        },
        "description": "Native immerle endpoints (first-run setup, capability\ndiscovery, friends, activity feed, collaborative playlists and\nsynchronized Jam sessions) that complement the Subsonic /\nOpenSubsonic API served under /rest/.\n\nAuthenticated endpoints reuse Subsonic credentials passed as\nquery parameters: u (username) plus either p (password) or the\ntoken pair t (md5(password+salt)) and s (salt), and c (client\nname). Setup and capability endpoints are unauthenticated.",
        "license": {
            "name": "See repository"
        },
        "title": "immerle extension API",
        "version": "1.0.0"
    },
...
    "paths": {
        "/account": {
            "get": {
                "description": "Reads (GET) or updates (POST) the authenticated user's own account. POST is a partial update — only fields present are changed. Lets a user set their display name and email themselves.",
                "parameters": [
                    {
                        "description": "Subsonic username (or use a Bearer token)",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "POST only: free-text UI name (empty clears it)",
                        "in": "query",
                        "name": "displayName",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "POST only: email address (empty clears it)",
                        "in": "query",
                        "name": "email",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.AccountResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    }
                },
                "summary": "Get or update your account",
                "tags": [
                    "users"
                ]
            },
            "post": {
                "description": "Reads (GET) or updates (POST) the authenticated user's own account. POST is a partial update — only fields present are changed. Lets a user set their display name and email themselves.",
                "parameters": [
                    {
                        "description": "Subsonic username (or use a Bearer token)",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "POST only: free-text UI name (empty clears it)",
                        "in": "query",
                        "name": "displayName",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "POST only: email address (empty clears it)",
                        "in": "query",
                        "name": "email",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.AccountResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    }
                },
                "summary": "Get or update your account",
                "tags": [
                    "users"
                ]
            }
        },
...
        "/admin/cleanup": {
            "get": {
                "description": "Admin only. GET reports the eviction sweep state; POST with enabled=true|false turns the background sweep on or off at runtime (persisted; hot).",
                "parameters": [
                    {
                        "description": "Subsonic username (or use a Bearer token)",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "POST only: enable or disable the sweep",
                        "in": "query",
                        "name": "enabled",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.CleanupStatusResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    }
                },
                "summary": "Get or toggle the cleanup sweep",
                "tags": [
                    "admin"
                ]
            },
            "post": {
                "description": "Admin only. GET reports the eviction sweep state; POST with enabled=true|false turns the background sweep on or off at runtime (persisted; hot).",
                "parameters": [
                    {
                        "description": "Subsonic username (or use a Bearer token)",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "POST only: enable or disable the sweep",
                        "in": "query",
                        "name": "enabled",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.CleanupStatusResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    }
                },
                "summary": "Get or toggle the cleanup sweep",
                "tags": [
                    "admin"
                ]
            }
        },
        "/admin/cleanup/run": {
            "post": {
                "description": "Admin only. Runs one eviction pass immediately and returns how many provider downloads were removed. Works even when the background sweep is disabled.",
                "parameters": [
                    {
                        "description": "Subsonic username (or use a Bearer token)",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.CleanupRunResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    }
                },
                "summary": "Run the cleanup sweep now",
                "tags": [
                    "admin"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
...
}